mv duplicate_pngs_from_folder to general_utils
